+++ tests/testthat/test-hc-sims.R
@@ -1,6 +1,6 @@
 test_that("hc_sims 1 sim", {
   withr::with_seed(42, {
-    sims <- ssd_simulate_data("rlnorm", seed = 10, nsim = 1L)
+    sims <- ssd_sim_data("rlnorm", seed = 10, nsim = 1L)
     sims <- ssd_fit_dists_sims(sims, seed = 10)
     sims <- ssd_hc_sims(sims)
   })
@@ -10,7 +10,7 @@ test_that("hc_sims 1 sim", {
 
 test_that("hc_sims 1 sim ci", {
   withr::with_seed(42, {
-    sims <- ssd_simulate_data("rlnorm", seed = 10, nsim = 1L)
+    sims <- ssd_sim_data("rlnorm", seed = 10, nsim = 1L)
     sims <- ssd_fit_dists_sims(sims, seed = 10)
     sims <- ssd_hc_sims(sims, ci = TRUE)
   })