+++ tests/testthat/test-simulate-data.R
@@ -1,70 +1,70 @@
-test_that("ssd_simulate_data.function works", {
+test_that("ssd_sim_data.function works", {
   withr::local_seed(42)
-  data <- ssd_simulate_data(ssdtools::ssd_rlnorm, nrow = 5, nsim = 10)
+  data <- ssd_sim_data(ssdtools::ssd_rlnorm, nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "function_unnest")
 })
 
-test_that("ssd_simulate_data.function vectorized", {
+test_that("ssd_sim_data.function vectorized", {
   withr::local_seed(42)
-  data <- ssd_simulate_data(ssdtools::ssd_rlnorm, nrow = c(5, 10), nsim = 2)
+  data <- ssd_sim_data(ssdtools::ssd_rlnorm, nrow = c(5, 10), nsim = 2)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "function_vectorize")
 })
 
-test_that("ssd_simulate_data.character works", {
+test_that("ssd_sim_data.character works", {
   withr::local_seed(42)
-  data <- ssd_simulate_data("rlnorm", nrow = 5, nsim = 10)
+  data <- ssd_sim_data("rlnorm", nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "character_unnest")
 })
 
-test_that("ssd_simulate_data.character vectorized", {
+test_that("ssd_sim_data.character vectorized", {
   withr::local_seed(42)
-  data <- ssd_simulate_data("rlnorm", nrow = c(5, 10), nsim = 2)
+  data <- ssd_sim_data("rlnorm", nrow = c(5, 10), nsim = 2)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "character_vectorize")
 })
 
-test_that("ssd_simulate_data.fitdists works top", {
+test_that("ssd_sim_data.fitdists works top", {
   data <- ssddata::ccme_boron
   fit <- ssdtools::ssd_fit_dists(data, dists = c("lnorm", "gamma"))
   withr::local_seed(42)
-  data <- ssd_simulate_data(fit, dist_sim = "lnorm", nrow = 5, nsim = 10)
+  data <- ssd_sim_data(fit, dist_sim = "lnorm", nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "fitdists_top_unnest")
 })
 
-test_that("ssd_simulate_data.fitdists works multi", {
+test_that("ssd_sim_data.fitdists works multi", {
   data <- ssddata::ccme_boron
   fit <- ssdtools::ssd_fit_dists(data, dists = c("lnorm", "gamma"))
   withr::local_seed(42)
-  data <- ssd_simulate_data(fit, dist_sim = "multi", nrow = 5, nsim = 10)
+  data <- ssd_sim_data(fit, dist_sim = "multi", nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "fitdists_multi_unnest")
 })
 
-test_that("ssd_simulate_data.fitdists works name", {
+test_that("ssd_sim_data.fitdists works name", {
   data <- ssddata::ccme_boron
   fit <- ssdtools::ssd_fit_dists(data, dists = c("lnorm", "gamma"))
   withr::local_seed(42)
-  data <- ssd_simulate_data(fit, dist_sim = "gamma", nrow = 5, nsim = 10)
+  data <- ssd_sim_data(fit, dist_sim = "gamma", nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "fitdists_name_unnest")
 })
 
-test_that("ssd_simulate_data.fitdists works vectorized", {
+test_that("ssd_sim_data.fitdists works vectorized", {
   data <- ssddata::ccme_boron
   fit <- ssdtools::ssd_fit_dists(data, dists = c("lnorm", "gamma"))
   withr::local_seed(42)
-  data <- ssd_simulate_data(
+  data <- ssd_sim_data(
     fit,
     dist_sim = c("top", "multi", "lnorm"),
     nrow = c(5, 10),
@@ -75,11 +75,11 @@ test_that("ssd_simulate_data.fitdists works vectorized", {
   expect_snapshot_data(data, "fitdists_name_vectorize")
 })
 
-test_that("ssd_simulate_data.fitdists works all vectorized", {
+test_that("ssd_sim_data.fitdists works all vectorized", {
   data <- ssddata::ccme_boron
   fit <- ssdtools::ssd_fit_dists(data, dists = c("lnorm", "gamma"))
   withr::local_seed(42)
-  data <- ssd_simulate_data(
+  data <- ssd_sim_data(
     fit,
     dist_sim = c("multi", "lnorm", "all"),
     nrow = c(5, 10),
@@ -90,27 +90,27 @@ test_that("ssd_simulate_data.fitdists works all vectorized", {
   expect_snapshot_data(data, "fitdists_name_all_vectorize")
 })
 
-test_that("ssd_simulate_data.tmbfit works", {
+test_that("ssd_sim_data.tmbfit works", {
   data <- ssddata::ccme_boron
   tmbfit <- ssdtools::ssd_fit_dists(data, dists = "lnorm")[[1]]
   withr::local_seed(42)
-  data <- ssd_simulate_data(tmbfit, nrow = 5, nsim = 10)
+  data <- ssd_sim_data(tmbfit, nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "tmbfit_unnest")
 })
 
-test_that("ssd_simulate_data.data.frame works", {
+test_that("ssd_sim_data.data.frame works", {
   withr::local_seed(42)
-  data <- ssd_simulate_data(ssddata::ccme_boron, nrow = 5, nsim = 10)
+  data <- ssd_sim_data(ssddata::ccme_boron, nrow = 5, nsim = 10)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "data_frame_unnest")
 })
 
-test_that("ssd_simulate_data.data.frame vectorized", {
+test_that("ssd_sim_data.data.frame vectorized", {
   withr::local_seed(42)
-  data <- ssd_simulate_data(ssddata::ccme_boron, nrow = c(5, 10), nsim = 2)
+  data <- ssd_sim_data(ssddata::ccme_boron, nrow = c(5, 10), nsim = 2)
   expect_snapshot(data)
   data <- tidyr::unnest(data, cols = c(data))
   expect_snapshot_data(data, "data_frame_vectorize")